requirements.txt
django-extensions==3.2.3
django-localflavor==4.0
django-ninja==0.22.2
django-widget-tweaks==1.4.12
Django==4.2.3
python-decouple==3.8
isort==5.12.0
autopep8==2.0.2
djhtml==3.0.6
